requirements.txt
numpy
imageio
trimesh
opencv-python
pillow
termcolor
open3d
torch
cython
fvcore
chumpy==0.70
plyfile
scikit-image==0.18.1
transforms3d==0.3.1
transformers==4.20.1
jax==0.3.14
jaxlib==0.3.14
opendr==0.78
GitPython
tensorboard~=2.9.1
tqdm
termcolor
manotorch@git+https://github.com/lixiny/manotorch.git
pytorch3d@git+https://github.com/facebookresearch/pytorch3d@v0.7.1
